Migrate the rest of the tabs commands into excmd_tabs
